Cache api_keys (#93)
* Cache api_keys

Currently have request to an internal service requires at least one
DB hit. For some services this is noise, for others it's the only
reason they have a databse. In a few cases, API key look up is
what the database is spending most of its time on.

Since this data is effectively immutable it's a great candidate for
caching. All apps would benefit from this.

This caches API keys in memory using a bounded LRU cache.

* Run Cache injector at define time

* Allow configuration via ENV

* Add new configuration values for api key cache

* migrate cache injector to a cache wrapper

* Use new cache key wrapper for api client wrapper

* Refactor api client access for configurable cache

* Update generators to include the new config values

* Test cache and db paths

* update docs

* Speling

* Disable clear when cache is disabled

* An RC

* More docs

* remove deprecated classes and upgrade mechanism (#94)

Co-authored-by: David Copeland <22282+davetron5000@users.noreply.github.com>