Add disabled at (#101)
* Backfill tests

Structure the code slightly differently in order to facilitate testing,
and then test with a database and a full Rails stack for better
assurance we have things working correctly.

* Add usage of disabled_at to allow old keys

When old keys are used, and assuming the disabled_at field is set, this
allows those old keys to be successful, but warns that the key is
disabled and will be removed soon.

Use defaults for 6.0 so that test work for Rails 6 runner

Update the generators to reduce whitespace

Blank lines were generated by the generators in a few places, which
cause some style format validators to fail, such as rubocop.

* Update version to 4.2.0

Note that we are moving from 4.0.2 to 4.2.0. This is because there were
a set of RC builds and tags for 4.1.0 which never made it to a 4.1.0
release. Nevertheless, it seems prudent to move to a new minor revision.